+++ flutter_quill_extensions/lib/src/common/utils/web/web_stub.dart
@@ -1,0 +1,19 @@
+void createHtmlImageElement({
+  required String src,
+  required String height,
+  required String width,
+  required String margin,
+  required String alignSelf,
+}) =>
+    throw UnimplementedError(
+        'A stub method is called, createHtmlImageElement is for web platforms only.');
+
+void createHtmlIFrameElement({
+  required String src,
+  required String height,
+  required String width,
+  required String margin,
+  required String alignSelf,
+}) =>
+    throw UnimplementedError(
+        'A stub method is called, createHtmlIFrameElement is for web platforms only.');