+++ flutter_quill_extensions/lib/src/editor/video/video_web_embed.dart
@@ -1,12 +1,9 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_quill/flutter_quill.dart';
-import 'package:universal_html/html.dart' as html;
 
-import '../../common/utils/dart_ui/dart_ui_fake.dart'
-    if (dart.library.js_interop) '../../common/utils/dart_ui/dart_ui_real.dart'
-    as ui;
 import '../../common/utils/element_utils/element_web_utils.dart';
 import '../../common/utils/utils.dart';
+import '../../common/utils/web/web.dart';
 import 'config/video_web_config.dart';
 import 'youtube_video_url.dart';
 
@@ -40,16 +37,12 @@ class QuillEditorWebVideoEmbedBuilder extends EmbedBuilder {
     final (height, width, margin, alignment) =
         getWebElementAttributes(embedContext.node);
 
-    ui.PlatformViewRegistry().registerViewFactory(
-      videoUrl,
-      (id) => html.IFrameElement()
-        ..style.width = width
-        ..style.height = height
-        ..src = videoUrl
-        ..style.border = 'none'
-        ..style.margin = margin
-        ..style.alignSelf = alignment
-        ..attributes['loading'] = 'lazy',
+    createHtmlIFrameElement(
+      src: videoUrl,
+      width: width,
+      height: height,
+      margin: margin,
+      alignSelf: alignment,
     );
 
     return SizedBox(