+++ lib/src/editor/config/editor_config.dart
@@ -1,7 +1,6 @@
 import 'dart:ui' as ui;
 
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart' show experimental;
 
@@ -73,7 +72,6 @@ class QuillEditorConfig {
     this.contextMenuBuilder,
     this.editorKey,
     this.requestKeyboardFocusOnCheckListChanged = false,
-    @experimental this.magnifierConfiguration,
     this.textInputAction = TextInputAction.newline,
     this.enableScribble = false,
     this.onScribbleActivated,
@@ -338,12 +336,11 @@ class QuillEditorConfig {
 
   // Returns whether gesture is handled
   final bool Function(
-          TapDragDownDetails details, TextPosition Function(Offset offset))?
-      onTapDown;
+      TapDownDetails details, TextPosition Function(Offset offset))? onTapDown;
 
   // Returns whether gesture is handled
   final bool Function(
-      TapDragUpDetails details, TextPosition Function(Offset offset))? onTapUp;
+      TapUpDetails details, TextPosition Function(Offset offset))? onTapUp;
 
   // Returns whether gesture is handled
   final bool Function(
@@ -436,11 +433,6 @@ class QuillEditorConfig {
   /// should we request keyboard focus??
   final bool requestKeyboardFocusOnCheckListChanged;
 
-  /// This feature is currently experimental and only supported
-  /// on **Android** and **iOS**.
-  @experimental
-  final TextMagnifierConfiguration? magnifierConfiguration;
-
   /// Default to [TextInputAction.newline]
   final TextInputAction textInputAction;
 
@@ -488,11 +480,9 @@ class QuillEditorConfig {
     Brightness? keyboardAppearance,
     ScrollPhysics? scrollPhysics,
     ValueChanged<String>? onLaunchUrl,
-    bool Function(
-            TapDragDownDetails details, TextPosition Function(Offset offset))?
+    bool Function(TapDownDetails details, TextPosition Function(Offset offset))?
         onTapDown,
-    bool Function(
-            TapDragUpDetails details, TextPosition Function(Offset offset))?
+    bool Function(TapUpDetails details, TextPosition Function(Offset offset))?
         onTapUp,
     Iterable<EmbedBuilder>? embedBuilders,
     EmbedBuilder? unknownEmbedBuilder,
@@ -512,7 +502,6 @@ class QuillEditorConfig {
     GlobalKey<EditorState>? editorKey,
     TextSelectionThemeData? textSelectionThemeData,
     bool? requestKeyboardFocusOnCheckListChanged,
-    TextMagnifierConfiguration? magnifierConfiguration,
     TextInputAction? textInputAction,
     bool? enableScribble,
     void Function()? onScribbleActivated,
@@ -581,8 +570,6 @@ class QuillEditorConfig {
       requestKeyboardFocusOnCheckListChanged:
           requestKeyboardFocusOnCheckListChanged ??
               this.requestKeyboardFocusOnCheckListChanged,
-      magnifierConfiguration:
-          magnifierConfiguration ?? this.magnifierConfiguration,
       textInputAction: textInputAction ?? this.textInputAction,
       enableScribble: enableScribble ?? this.enableScribble,
       onScribbleActivated: onScribbleActivated ?? this.onScribbleActivated,